Simpler library usage:
* Renamed OpenTelemetry to opentelemetry.tracing.Tracing: to be able to deal with extended functionality of metrics in the future (under opentelemetry.metrics)
* moved all methods into main object: Tracing. Allows for simpler usage of the library
