+++ modules/opentelemetry/src/main/scala/zio/opentelemetry/tracing/attributevalue/AttributeValueConverter.scala
@@ -1,4 +1,4 @@
-package zio.opentelemetry.attributevalue
+package zio.opentelemetry.tracing.attributevalue
 
 import io.opentelemetry.common.AttributeValue
